+++ apstools/plans/labels_to_streams.py
@@ -96,13 +96,15 @@ from bluesky.magics import get_labeled_devices
 from bluesky.utils import make_decorator
 from bluesky.utils import single_gen
 from deprecated.sphinx import versionadded
+from deprecated.sphinx import versionchanged
 
 from ..utils import getDefaultNamespace
 from .doc_run import write_stream
 
 
+@versionchanged(version="1.7.6", reason="Caller can specify namespace dict.")
 @versionadded(version="1.6.11")
-def label_stream_stub(labels=None, fmt=None, bec=None):
+def label_stream_stub(labels=None, fmt=None, bec=None, ns: dict[str, object] = None):
     """
     Writes ophyd-labeled objects to open bluesky run streams. One stream per label.
 
@@ -122,11 +124,16 @@ def label_stream_stub(labels=None, fmt=None, bec=None):
         *obj*:
         Instance of bluesky BestEffortCallback.
         Default: selected from default namespace, if available.
+
+    ns
+        *dict*:
+        Namespace dictionary to search for labeled objects.
+        Default: selected from default namespace, if available.
     """
     from bluesky.callbacks.best_effort import BestEffortCallback
 
     fmt = fmt or "label_{}"
-    ns = getDefaultNamespace()
+    ns = ns or getDefaultNamespace()
     devices = get_labeled_devices(ns)
     labels = labels or list(devices.keys())
     if not isinstance(labels, (list, tuple)):
@@ -158,8 +165,9 @@ class When(Enum):
     END = "end"
 
 
+@versionchanged(version="1.7.6", reason="Caller can specify namespace dict.")
 @versionadded(version="1.6.11")
-def label_stream_wrapper(plan, labels, fmt=None, when="start"):
+def label_stream_wrapper(plan, labels, fmt=None, when="start", ns: dict[str, object] = None):
     """
     Decorator support: Write labeled device(s) to stream(s).  Either at "start" or "end".
 
@@ -171,7 +179,7 @@ def label_stream_wrapper(plan, labels, fmt=None, when="start"):
     labels
         *[str]* (or *str*):
         List of configured ophyd object "labels".
-        Passed through to :meth:`~apstools.plans.write_label_stream()`.
+        Passed through to :meth:`~apstools.plans.label_stream_stub()`.
         Default: ``None`` (meaning all).
     fmt
         *str*:
@@ -193,6 +201,11 @@ def label_stream_wrapper(plan, labels, fmt=None, when="start"):
         The ``str`` value can be expressed in either upper or lower case.
 
         Default: ``"start"``
+
+    ns
+        *dict*:
+        Namespace dictionary to search for labeled objects.
+        Default: selected from default namespace, if available.
     """
     try:
         if isinstance(when, str):
@@ -207,7 +220,7 @@ def label_stream_wrapper(plan, labels, fmt=None, when="start"):
         if msg.command == "open_run":
 
             def new_gen():
-                yield from label_stream_stub(labels, fmt=fmt)
+                yield from label_stream_stub(labels, fmt=fmt, ns=ns)
 
             return single_gen(msg), new_gen()
         else:
@@ -217,7 +230,7 @@ def label_stream_wrapper(plan, labels, fmt=None, when="start"):
         if msg.command == "close_run":
 
             def new_gen():
-                yield from label_stream_stub(labels, fmt=fmt)
+                yield from label_stream_stub(labels, fmt=fmt, ns=ns)
                 yield msg
 
             return new_gen(), None